Home3
Update of Home.html. Replaced the lorum ipsum text on the main body of the page with explaining the business based on the details given to us by the client. Additionally, I updated the formatting and grammar of the information panel on the left. (Something that also shows in the Prices.html update when I post it in a few minutes)
 
We should be finished with this, content wise. Most of our work from hereon likely will revolve around editing the styling sheet, and making it look nice/presentable.