Prices2
Updated the file to remove the filler text in the main body. Also updated the information on the side, re-formatting some of the text to look nicer, as well as the bottom information bar.

I would personally recommend that we use the same image on the home page on this page as well, so the look of the website is consistent. Currently, if we click on different pages, the general look is noticeably different. Let me know what you think.

Beyond that, the content portion of this page is likely finished. Further work will likely come from the style sheet.