Prices.html
Beyond the base file that we've already established, this is posting it to the master branch, so when we merge various changes, we can compare.

Also, it should be noted that in this page, there is a slight alignment issue. I will look into amending this so the look of it is aesthetic. I will be uploading my version of the CSS, after this file is finished uploading.