StyleSheet.css
Many of the changes here revolve around how the website looks. Some of the colors are adjusted, so it isn't flat colors. I also attempted to create constant sizes for some of the blocks on the website, so the look is more consistent between the two different pages.

As I mentioned within the prices.html file, there is currently an alignment issue that I'm trying to resolve, and would appreciate help if possible. Specifically, the body on the prices page is just slightly longer than the aside, and it messses up border alignment as a result.