StyleSheet.css
With these edits, the main alignment issue with the pages has been solved, and are now operating at a consistent look to the website. To be more specific as to what was changed, I created separate IDs for the main bodies on each page, so we could change them individually, as we needed.

Given that the error has been fixed, I'm personally saying that this project is finished, design-wise. Please, using the Home.html and Price.html files with this one, go over the website and give the okay on it if there's nothing else we should do, so I can submit the files to the client and get his 'okay' on it.